requirements.txt
Django>=3.0,<3.1
wagtail>=2.8,<2.9
mysqlclient>=1.4.6
pymysql>=0.9.3
djangorestframework>=2.29
markdown>=3.1.1
django-filter>=2.2.0
django-rest-auth>=0.9.5
requests>=2.22.0
python-dateutil>=2.8.1